refactor(field-trip): add field trip actions; clean up editable section
