+++ lib/components/admin/call-taker-controls.js
@@ -1,9 +1,10 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 
+import * as apiActions from '../../actions/api'
 import * as callTakerActions from '../../actions/call-taker'
-import { routingQuery } from '../../actions/api'
-import { setMainPanelContent } from '../../actions/ui'
+import * as fieldTripActions from '../../actions/field-trip'
+import * as uiActions from '../../actions/ui'
 import CallTimeCounter from './call-time-counter'
 import Icon from '../narrative/icon'
 
@@ -169,24 +170,15 @@ const mapStateToProps = (state, ownProps) => {
   }
 }
 
-const {
-  beginCall,
-  endCall,
-  fetchCalls,
-  fetchFieldTrips,
-  toggleCallHistory,
-  toggleFieldTrips
-} = callTakerActions
-
 const mapDispatchToProps = {
-  beginCall,
-  endCall,
-  fetchCalls,
-  fetchFieldTrips,
-  routingQuery,
-  setMainPanelContent,
-  toggleCallHistory,
-  toggleFieldTrips
+  beginCall: callTakerActions.beginCall,
+  endCall: callTakerActions.endCall,
+  fetchCalls: callTakerActions.fetchCalls,
+  fetchFieldTrips: fieldTripActions.fetchFieldTrips,
+  routingQuery: apiActions.routingQuery,
+  setMainPanelContent: uiActions.setMainPanelContent,
+  toggleCallHistory: callTakerActions.toggleCallHistory,
+  toggleFieldTrips: fieldTripActions.toggleFieldTrips
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(CallTakerControls)
