+++ lib/components/admin/editable-section.js
@@ -6,6 +6,11 @@ import {
   Val
 } from './styled'
 
+/**
+ * This component will render a set of fields along with a 'Change/Edit' button
+ * that toggles an editable state. In its editable state, the fields/values are
+ * rendered as form inputs, which can be edited and then saved/persisted.
+ */
 export default class EditableSection extends Component {
   state = {
     data: {},
@@ -16,7 +21,9 @@ export default class EditableSection extends Component {
 
   _getVal = (fieldName) => this._exists(this.state.data[fieldName])
     ? this.state.data[fieldName]
-    : this.props.request[fieldName]
+    : this._exists(this.props.request[fieldName])
+      ? this.props.request[fieldName]
+      : ''
 
   _onChange = (fieldName, value) => {
     const stateUpdate = this.state.data
@@ -43,41 +50,39 @@ export default class EditableSection extends Component {
   }
 
   render () {
-    const {fields, header, inputStyle, request, valueFirst} = this.props
+    const {fields, children, inputStyle, request, valueFirst} = this.props
     const {isEditing} = this.state
     if (!request) return null
     return (
       <>
-        <P>
-          {header}
-          <span className='pull-right' style={{marginRight: '15px'}}>
-            {!isEditing
-              ? <Button
+        {children}
+        <span className='pull-right' style={{marginRight: '15px'}}>
+          {!isEditing
+            ? <Button
+              bsSize='xsmall'
+              onClick={this._toggleEditing}
+            >
+              Change
+            </Button>
+            : <>
+              <Button
                 bsSize='xsmall'
+                bsStyle='link'
                 onClick={this._toggleEditing}
               >
-                Change
+                Cancel
               </Button>
-              : <>
-                <Button
-                  bsSize='xsmall'
-                  bsStyle='link'
-                  onClick={this._toggleEditing}
-                >
-                  Cancel
-                </Button>
-                <Button
-                  bsSize='xsmall'
-                  bsStyle='link'
-                  disabled={Object.keys(this.state.data).length === 0}
-                  onClick={this._onClickSave}
-                >
-                  Save
-                </Button>
-              </>
-            }
-          </span>
-        </P>
+              <Button
+                bsSize='xsmall'
+                bsStyle='link'
+                disabled={Object.keys(this.state.data).length === 0}
+                onClick={this._onClickSave}
+              >
+                Save
+              </Button>
+            </>
+          }
+        </span>
         {fields.map(f => {
           const input = (
             <InputToggle
@@ -103,6 +108,11 @@ export default class EditableSection extends Component {
   }
 }
 
+/**
+ * This component renders either the specified value for a given field or, if
+ * in an active editing state, the associated field's respective input (e.g.
+ * text, number, or select).
+ */
 class InputToggle extends Component {
   _onChange = (evt) => {
     const {fieldName, inputProps = {}, onChange} = this.props
