+++ lib/components/admin/field-trip-details.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 import { DropdownButton, MenuItem } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import * as callTakerActions from '../../actions/call-taker'
+import * as fieldTripActions from '../../actions/field-trip'
 import DraggableWindow from './draggable-window'
 import EditableSection from './editable-section'
 import FieldTripNotes from './field-trip-notes'
@@ -13,10 +13,12 @@ import {
   B,
   Button,
   Container,
-  Full,
+  FullWithMargin,
   Half,
   Header,
-  P
+  InlineHeader,
+  P,
+  Text
 } from './styled'
 import TripStatus from './trip-status'
 import Updatable from './updatable'
@@ -59,6 +61,9 @@ const PAYMENT_FIELDS = [
  */
 class FieldTripDetails extends Component {
   _editSubmitterNotes = (val) => this.props.editSubmitterNotes(this.props.request, val)
+
+  _getRequestLink = (path, isPublic = false) =>
+    `${this.props.datastoreUrl}/${isPublic ? 'public/' : ''}fieldtrip/${path}?requestId=${this.props.request.id}`
 
   _onCloseActiveFieldTrip = () => this.props.setActiveFieldTrip(null)
 
@@ -112,10 +117,16 @@ class FieldTripDetails extends Component {
               id='field-trip-menu'
               title='View'
             >
-              <MenuItem onClick={this._showRouteViewer}>
+              <MenuItem
+                href={this._getRequestLink('feedbackForm', true)}
+                target='_blank'
+              >
                 <Icon type='link' /> Feedback link
               </MenuItem>
-              <MenuItem onClick={this._startOver}>
+              <MenuItem
+                href={this._getRequestLink('receipt')}
+                target='_blank'
+              >
                 <Icon type='file-text-o' /> Receipt link
               </MenuItem>
             </DropdownButton>
@@ -160,8 +171,8 @@ class FieldTripDetails extends Component {
           </Half>
           <Half>
             <EditableSection
+              children={<Text><B>{total}</B> total group size</Text>}
               fields={GROUP_FIELDS}
-              header={<span><B>{total}</B> total group size</span>}
               inputStyle={{lineHeight: '0.8em', padding: '0px', width: '50px'}}
               onChange={setRequestGroupSize}
               request={request}
@@ -173,19 +184,20 @@ class FieldTripDetails extends Component {
             request={request}
           />
           <TripStatus request={request} />
-          <Full>
+          <FullWithMargin>
             <EditableSection
-              fields={PAYMENT_FIELDS}
-              header={
-                <Header>
+              // Use inline header so that 'Change' button appears on top line.
+              children={
+                <InlineHeader>
                   <Icon type='credit-card' /> Payment information
-                </Header>
+                </InlineHeader>
               }
+              fields={PAYMENT_FIELDS}
               inputStyle={{lineHeight: '0.8em', padding: '0px', width: '100px'}}
               onChange={setRequestPaymentInfo}
               request={request}
             />
-          </Full>
+          </FullWithMargin>
           <FieldTripNotes
             addFieldTripNote={addFieldTripNote}
             deleteFieldTripNote={deleteFieldTripNote}
@@ -203,22 +215,23 @@ const mapStateToProps = (state, ownProps) => {
   return {
     callTaker: state.callTaker,
     currentQuery: state.otp.currentQuery,
+    datastoreUrl: state.otp.config.datastoreUrl,
     dateFormat: getDateFormat(state.otp.config),
     request
   }
 }
 
 const mapDispatchToProps = {
-  addFieldTripNote: callTakerActions.addFieldTripNote,
-  deleteFieldTripNote: callTakerActions.deleteFieldTripNote,
-  editSubmitterNotes: callTakerActions.editSubmitterNotes,
-  fetchQueries: callTakerActions.fetchQueries,
-  setActiveFieldTrip: callTakerActions.setActiveFieldTrip,
-  setFieldTripFilter: callTakerActions.setFieldTripFilter,
-  setRequestGroupSize: callTakerActions.setRequestGroupSize,
-  setRequestPaymentInfo: callTakerActions.setRequestPaymentInfo,
-  setRequestStatus: callTakerActions.setRequestStatus,
-  toggleFieldTrips: callTakerActions.toggleFieldTrips
+  addFieldTripNote: fieldTripActions.addFieldTripNote,
+  deleteFieldTripNote: fieldTripActions.deleteFieldTripNote,
+  editSubmitterNotes: fieldTripActions.editSubmitterNotes,
+  fetchQueries: fieldTripActions.fetchQueries,
+  setActiveFieldTrip: fieldTripActions.setActiveFieldTrip,
+  setFieldTripFilter: fieldTripActions.setFieldTripFilter,
+  setRequestGroupSize: fieldTripActions.setRequestGroupSize,
+  setRequestPaymentInfo: fieldTripActions.setRequestPaymentInfo,
+  setRequestStatus: fieldTripActions.setRequestStatus,
+  toggleFieldTrips: fieldTripActions.toggleFieldTrips
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(FieldTripDetails)
