+++ lib/components/admin/field-trip-windows.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 import { Badge } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import * as callTakerActions from '../../actions/call-taker'
+import * as fieldTripActions from '../../actions/field-trip'
 import DraggableWindow from './draggable-window'
 import FieldTripDetails from './field-trip-details'
 import Icon from '../narrative/icon'
@@ -235,11 +235,11 @@ const mapStateToProps = (state, ownProps) => {
 }
 
 const mapDispatchToProps = {
-  fetchFieldTripDetails: callTakerActions.fetchFieldTripDetails,
-  fetchFieldTrips: callTakerActions.fetchFieldTrips,
-  setActiveFieldTrip: callTakerActions.setActiveFieldTrip,
-  setFieldTripFilter: callTakerActions.setFieldTripFilter,
-  toggleFieldTrips: callTakerActions.toggleFieldTrips
+  fetchFieldTripDetails: fieldTripActions.fetchFieldTripDetails,
+  fetchFieldTrips: fieldTripActions.fetchFieldTrips,
+  setActiveFieldTrip: fieldTripActions.setActiveFieldTrip,
+  setFieldTripFilter: fieldTripActions.setFieldTripFilter,
+  toggleFieldTrips: fieldTripActions.toggleFieldTrips
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(FieldTripWindows)
