+++ lib/components/admin/styled.js
@@ -1,5 +1,5 @@
 import { Button as BsButton } from 'react-bootstrap'
-import styled from 'styled-components'
+import styled, {css} from 'styled-components'
 
 export const B = styled.strong``
 
@@ -20,14 +20,30 @@ export const Full = styled.div`
   width: 100%
 `
 
+export const FullWithMargin = styled(Full)`
+  margin-top: 10px;
+`
+
 export const Header = styled.h4`
   margin-bottom: 5px;
   width: 100%;
 `
 
-export const P = styled.p`
+export const InlineHeader = styled(Header)`
+  display: inline;
+`
+
+export const textCss = css`
   font-size: 0.9em;
   margin-bottom: 0px;
+`
+
+export const P = styled.p`
+  ${textCss}
+`
+
+export const Text = styled.span`
+  ${textCss}
 `
 
 export const Val = styled.span`
