+++ lib/components/admin/trip-status.js
@@ -3,7 +3,7 @@ import moment from 'moment'
 import React, {Component} from 'react'
 import { connect } from 'react-redux'
 
-import * as callTakerActions from '../../actions/call-taker'
+import * as fieldTripActions from '../../actions/field-trip'
 import * as formActions from '../../actions/form'
 import Icon from '../narrative/icon'
 import {
@@ -101,8 +101,8 @@ const mapStateToProps = (state, ownProps) => {
 }
 
 const mapDispatchToProps = {
-  planTrip: callTakerActions.planTrip,
-  saveRequestTrip: callTakerActions.saveRequestTrip,
+  planTrip: fieldTripActions.planTrip,
+  saveRequestTrip: fieldTripActions.saveRequestTrip,
   setQueryParam: formActions.setQueryParam
 }
