+++ lib/components/app/call-taker-panel.js
@@ -5,7 +5,7 @@ import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
 import * as apiActions from '../../actions/api'
-import * as callTakerActions from '../../actions/call-taker'
+import * as fieldTripActions from '../../actions/field-trip'
 import * as formActions from '../../actions/form'
 import AddPlaceButton from '../form/add-place-button'
 import AdvancedOptions from '../form/call-taker/advanced-options'
@@ -289,7 +289,7 @@ const mapStateToProps = (state, ownProps) => {
 const mapDispatchToProps = {
   findRoutes: apiActions.findRoutes,
   routingQuery: apiActions.routingQuery,
-  setGroupSize: callTakerActions.setGroupSize,
+  setGroupSize: fieldTripActions.setGroupSize,
   setQueryParam: formActions.setQueryParam
 }
