+++ lib/util/call-taker.js
@@ -7,6 +7,17 @@ function placeToLatLonStr (place) {
 }
 
 /**
+ * @return {boolean} - whether a calltaker session is invalid
+ */
+export function sessionIsInvalid (session) {
+  if (!session || !session.sessionId) {
+    console.error('No valid OTP datastore session found.')
+    return true
+  }
+  return false
+}
+
+/**
  * Utility to map an OTP MOD UI search object to a Call Taker datastore query
  * object.
  */
